refactor: reorganize test suite for better maintainability
- Merge cell-formatting.test.ts into cells.test.ts
- Split manage.test.ts into focused files:
  - manage.test.ts: doc & sheet lifecycle only
  - worksheet-operations.test.ts: data manipulation operations
  - worksheet-features.test.ts: worksheet features (filters, formatting, metadata)
- Fix WorksheetDimensionProperties type to accept Partial for updateDimensionProperties

This reorganization makes the test suite more maintainable by grouping related tests together and splitting the large manage.test.ts file into logical units.
